requirements.txt

amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.11.3'
backports.zoneinfo[tzdata]==0.2.1; python_version < '3.9'
billiard==4.2.1; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
contourpy==1.1.1; python_version >= '3.8'
cycler==0.12.1; python_version >= '3.8'
django==4.2.16; python_version >= '3.8'
django-cors-headers==4.4.0; python_version >= '3.8'
django-extensions==3.2.3; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
fonttools==4.54.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
importlib-resources==6.4.5; python_version < '3.10'
iniconfig==2.0.0; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
kiwisolver==1.4.7; python_version >= '3.8'
kombu==5.4.2; python_version >= '3.8'
matplotlib==3.7.5; python_version >= '3.8'
numpy==1.24.4; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pandas==2.0.3; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
plotly==5.24.1; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyparsing==3.1.4; python_full_version >= '3.6.8'
pytest==8.3.3; python_version >= '3.8'
pytest-django==4.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
redis==5.1.1; python_version >= '3.8'
reportlab==4.2.5; python_version >= '3.7' and python_version < '4'
requests==2.32.3; python_version >= '3.8'
scikit-learn==1.3.2; python_version >= '3.8'
scipy==1.10.1; python_version < '3.12' and python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.1; python_version >= '3.8'
tenacity==9.0.0; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
tomli==2.0.2; python_version < '3.11'
typing-extensions==4.12.2; python_version < '3.11'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
zipp==3.20.2; python_version < '3.10'